+++ src/components/NightModeToggle.svelte
@@ -1,27 +1,11 @@
 <script lang="ts">
-  import { nightMode } from "../domain/store";
-
-  // add or remove dark mode stylesheet in a hacky way when nightMode state changes
-  $: {
-    let darkLink = document.getElementById("dark-theme");
-    if ($nightMode && !darkLink) {
-      darkLink = document.createElement("link");
-      darkLink.setAttribute("rel", "stylesheet");
-      darkLink.id = "dark-theme";
-      darkLink.setAttribute("href", "https://unpkg.com/bulma-prefers-dark");
-      document.head.appendChild(darkLink);
-    }
-    if (!$nightMode && darkLink) {
-      darkLink.remove();
-    }
-  }
+  import { nightMode } from "../domain/night-mode";
 </script>
 
-<main>
-  <div class="icon is-medium" on:click={nightMode.toggle}>
-    <i class="fa-lightbulb fa-lg fas" />
-  </div>
-</main>
-
-<style>
-</style>
+<span
+  class="icon is-medium"
+  on:click={nightMode.toggle}
+  data-tooltip="Night mode"
+>
+  <i class="fa-lightbulb fa-lg fas" />
+</span>
