+++ src/domain/User.ts
@@ -1,13 +1,12 @@
 import { writable } from "svelte/store";
 import * as waxjs from "@waxio/waxjs/dist";
 
-export const wax = new waxjs.WaxJS(
+export const wcw = new waxjs.WaxJS(
   "https://wax.greymass.com",
   undefined,
   undefined,
   false
 );
-
 export interface IAccountStore {
   account: string;
   loading: boolean;
@@ -19,56 +18,70 @@ export interface IAccountStore {
   client: waxjs.WaxJS;
 }
 
-const initial: IAccountStore = {
-  account: getFromSearch() || "",
-  loading: true,
-  client: wax,
-  isLoggedIn: false,
-};
-
 async function start(): Promise<void> {
-  console.log(store);
+  console.log(userStore);
   try {
-    store.update((state) => ({ ...state, loading: true }));
+    userStore.internal.update((state) => ({ ...state, loading: true }));
     const account = await autoLogin();
     setToSearch(account);
-    store.update((state) => ({ ...state, account, isLoggedIn: true }));
+    userStore.internal.update((state) => ({
+      ...state,
+      account,
+      isLoggedIn: true,
+    }));
   } catch (err) {
     console.warn(err);
   } finally {
-    store.update((state) => ({ ...state, loading: false }));
+    userStore.internal.update((state) => ({ ...state, loading: false }));
   }
 }
 
-export const store = writable<IAccountStore>(initial, (_set) => {
-  start();
-});
+function createUserStore() {
+  const internalStore = writable<IAccountStore>({
+    account: "",
+    loading: false,
+    client: wcw,
+    isLoggedIn: false,
+  });
+  return {
+    subscribe: internalStore.subscribe,
+    setAccount: (account: string) => {
+      internalStore.update((state) => ({ ...state, account }));
+    },
+    internal: internalStore,
+  };
+}
+
+export const userStore = createUserStore();
 
 //TODO logout???
 
 //checks if autologin is available and calls the normal login function if it is
 async function autoLogin(): Promise<string> {
-  const isAutoLoginAvailable = await wax.isAutoLoginAvailable();
+  const isAutoLoginAvailable = await wcw.isAutoLoginAvailable();
   if (!isAutoLoginAvailable) {
     throw new Error("cannot autologin");
   }
-  const account = await wax.login();
+  const account = await wcw.login();
   console.log("autologin worked", account);
   return account as string;
 }
 
 //normal login. Triggers a popup for non-whitelisted dapps
-export async function login(): Promise<void> {
+export async function wcwLogin(): Promise<void> {
   try {
-    store.update((state) => ({ ...state, loading: true }));
-    const account = (await wax.login()) as string;
-    store.update((state) => ({ ...state, account, isLoggedIn: true }));
-    setToSearch(account);
+    userStore.internal.update((state) => ({ ...state, loading: true }));
+    const account = (await wcw.login()) as string;
+    userStore.internal.update((state) => ({
+      ...state,
+      account,
+      isLoggedIn: true,
+    }));
   } catch (err) {
     console.log(err);
-    store.update((state) => ({ ...state, error: err }));
+    userStore.internal.update((state) => ({ ...state, error: err }));
   } finally {
-    store.update((state) => ({ ...state, loading: false }));
+    userStore.internal.update((state) => ({ ...state, loading: false }));
   }
 }
 
@@ -87,9 +100,4 @@ export function setToSearch(account: string): void {
   const search = new URLSearchParams(document.location.search);
   search.set("account", account);
   history.pushState(null, document.title, "?" + search.toString());
-}
-
-export function setAccount(account: string): void {
-  store.update((state) => ({ ...state, account }));
-  setToSearch(account);
 }
