--- src/domain/format.ts
@@ -1,7 +1,0 @@
-export function format(input: number | undefined) {
-  const numberToFormat = input || 0;
-  return numberToFormat.toLocaleString("en-us", {
-    minimumFractionDigits: 3,
-    maximumFractionDigits: 3,
-  });
-}
